# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PChat 的接口文档",
        "title": "PChat API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/users/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账户"
                ],
                "summary": "更新个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "占位符",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "common.StringValue": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        },
...
        "user.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.StringValue"
                        }
                    ]
                },
                "name": {
                    "description": "昵称",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.StringValue"
                        }
                    ]
                },
                "password": {
                    "description": "密码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.StringValue"
                        }
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "token": {
            "description": "jwt string",
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}